pyproject.toml
[project]
name = "google-adk-example"
version = "0.1.0"
description = "Minimal Google ADK example with google_search agent."
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "google-adk>=1.0.0",
  "python-dotenv>=1.0.0",
  "jinja2>=3.1.0",
  "pydantic>=2.0.0",
  "tomli>=2.0.0; python_version < '3.11'",
]

[tool.uv.sources]
google-adk = { path = "../../../study/ai-agents/google-adk-python", editable = true }

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "mypy",
  "fastapi",
  "litestar",
  "strawberry-graphql",
  "pydantic",
  "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["google_search_agent", "content_generator", "content_generator_agent"]

[tool.mypy]
strict = true
python_version = "3.10"
files = [
  "google_search_agent",
  "content_generator",
  "content_generator_agent",
]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "COM", # flake8-commas
  "EM", # flake8-errmsg
  "Q", # flake8-quotes
  "PTH", # flake8-use-pathlib
  "SIM", # flake8-simplify
  "TRY", # Trycertatops
  "PERF", # Perflint
  "RUF", # Ruff-specific rules
  "D", # pydocstyle
  "FA100",  # future annotations
]
ignore = [
  "COM812", # missing trailing comma, ruff format conflict
]
extend-safe-fixes = [
  "UP006",
  "UP007",
]
pyupgrade.keep-runtime-typing = false

[tool.ruff.lint.isort]
known-first-party = [
  "google_search_agent",
  "content_generator",
  "content_generator_agent",
]
combine-as-imports = true
required-imports = [
  "from __future__ import annotations",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"tests/**/*.py" = ["D100", "D102", "D103"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "--tb=short",
  "--no-header",
  "--showlocals",
]
asyncio_mode = "auto"
